+++ integration/js/pages/AppPage.js
@@ -65,7 +65,8 @@ function findAllElements() {
 
     eventReportingCheckBox: By.id('event-reporting'),
     eventReportingCheckBoxLabel: By.css('label[for="event-reporting"]'),
-    backgroundBlurFilterButton: By.id('dropdown-menu-filter-Background-Blur-10%-CPU'),
+    backgroundBlurFilterButton: By.id('dropdown-menu-filter-Background-Blur-40%-CPU'),
+    backgroundReplacementFilterButton: By.id('dropdown-menu-filter-Background-Replacement'),
     microphoneDropEchoButton: By.id('dropdown-menu-microphone-Echo'),
     echoReductionFeature: By.id('echo-reduction-capability'),
     echoReductionFeatureLabel: By.css('label[for="echo-reduction-capability"]'),
@@ -462,7 +463,7 @@ class AppPage {
     }
   }
 
-  async checkTranscriptsFromLastStart(expectedTranscriptContentBySpeaker, compareFn) {
+  async checkTranscriptsFromLastStart(expectedTranscriptContentBySpeaker, isMedicalTranscribe, compareFn) {
     const transcriptContainerText = await this.driver.findElement(elements.transcriptContainer).getText();
     const allTranscripts = transcriptContainerText.split('\n');
     if (allTranscripts.length < 1) {
@@ -504,8 +505,8 @@ class AppPage {
 
     for (let i = 0; i < actualSpeakers.length; i++) {
       const speaker = actualSpeakers[i];
-      if (!compareFn(actualTranscriptContentBySpeaker[speaker], expectedTranscriptContentBySpeaker[speaker])) {
-        console.log(`Transcript comparison failed, speaker ${speaker} actual content: "${actualTranscriptContentBySpeaker[speaker]}" does not match with expected: "${expectedTranscriptContentBySpeaker[speaker]}"`);
+      if (!compareFn(actualTranscriptContentBySpeaker[speaker], expectedTranscriptContentBySpeaker[speaker], isMedicalTranscribe)) {
+        console.log(`Transcript comparison failed, speaker: ${speaker} isMedicalTranscribe: ${isMedicalTranscribe} actual content: "${actualTranscriptContentBySpeaker[speaker]}" does not match with expected: "${expectedTranscriptContentBySpeaker[speaker]}"`);
         return false;
       }
     }
@@ -810,15 +811,33 @@ class AppPage {
   }
 
   async clickBackgroundBlurFilterFromDropDownMenu() {
-    await TestUtils.waitAround(1000);
-    const backgroundBlurButton = await this.driver.findElement(elements.backgroundBlurFilterButton);
-    await backgroundBlurButton.click();
+    await this.clickBackgroundFilterFromDropDownMenu(elements.backgroundBlurFilterButton);
   }
 
   async backgroundBlurCheck(attendeeId) {
-    await TestUtils.waitAround(4000);
     const expectedSumMin = 15805042;
     const expectedSumMax = 15940657;
+    return await this.backgroundFilterCheck(attendeeId, expectedSumMin, expectedSumMax);
+  }
+
+  async clickBackgroundReplacementFilterFromDropDownMenu() {
+    await this.clickBackgroundFilterFromDropDownMenu(elements.backgroundReplacementFilterButton);
+  }
+
+  async backgroundReplacementCheck(attendeeId) {
+    const expectedSumMin = 11000000;
+    const expectedSumMax = 11200000;
+    return await this.backgroundFilterCheck(attendeeId, expectedSumMin, expectedSumMax);
+  }
+
+  async clickBackgroundFilterFromDropDownMenu(buttonID) {
+    await TestUtils.waitAround(1000);
+    const backgroundBlurButton = await this.driver.findElement(buttonID);
+    await backgroundBlurButton.click();
+  }
+
+  async backgroundFilterCheck(attendeeId, expectedSumMin, expectedSumMax) {
+    await TestUtils.waitAround(4000);
     const videoElement = this.driver.findElement(By.xpath(`//*[contains(@class,'video-tile-nameplate') and contains(text(),'${attendeeId}')]`));
     const videoElementId = await videoElement.getAttribute('id');
     const seperatorIndex = videoElementId.lastIndexOf("-");
@@ -826,6 +845,7 @@ class AppPage {
       const tileIndex = parseInt(videoElementId.slice(seperatorIndex+1))
       if (tileIndex != NaN && tileIndex >= 0) {
         const videoImgSum = await this.driver.executeScript(this.getVideoImageSum(tileIndex));
+        console.log(`videoImgSum ${videoImgSum}`);
         if(videoImgSum < expectedSumMin || videoImgSum > expectedSumMax){
           console.log(`videoImgSum ${videoImgSum}`);
           return false;
@@ -834,6 +854,7 @@ class AppPage {
     }
     return true;
   }
+
 
   getVideoImageSum(videoId) {
     return "function getSum(total, num) {return total + num;};"