+++ src/videouplinkbandwidthpolicy/DefaultSimulcastUplinkPolicy.ts
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import Logger from '../logger/Logger';
-import AsyncScheduler from '../scheduler/AsyncScheduler';
 import SimulcastLayers from '../simulcastlayers/SimulcastLayers';
 import SimulcastTransceiverController from '../transceivercontroller/SimulcastTransceiverController';
 import { Maybe } from '../utils/Types';
@@ -133,8 +132,13 @@ export default class DefaultSimulcastUplinkPolicy implements SimulcastUplinkPoli
       if (this.shouldDisableSimulcast) {
         // See comment above `shouldDisableSimulcast` for usage.
         //
-        // The value of `newActiveStreams` is somewhat irrelevant since this single
-        // stream will adapt based on both sender and receiver network conditions.
+        // The value of `newActiveStreams` is somewhat irrelevant since in one to one calls
+        // we forward REMBs, so this single stream will adapt anywhere from < 100 kbps to 1200 kbps
+        // based on both sender and receiver network conditions. E.g. A receiver may calculate it's
+        // receive BWE as 300 kbps, send that in a REMB which is forwarded, and on receipt the sender
+        // will set its own BWE at 300 kbps, and start sending that as well (again, only for one-to-one
+        // calls). Additionally the value `kHi` is only relevant to the send side (via
+        // `encodingSimulcastLayersDidChange`) as it is not transmitted in anyform to the receiver.
         //
         // We use middle layer here to work around a bug in Chromium where
         // it seems when a transceiver is created when BWE is low (e.g. on a reconnection),
@@ -149,8 +153,9 @@ export default class DefaultSimulcastUplinkPolicy implements SimulcastUplinkPoli
         // and that technically the browser will still eventually try to send all 3 streams.
         //
         // Also note that due to some uninvestigated logic in bitrate allocation, Chromium
-        // will skip the bottom layer if we try setting it to 1200 kbps instead so it will still take a while to
-        // recover (as it needs to send padding until it reaches around 1000 kbps).
+        // will skip the bottom layer if we try setting it to 1200 kbps instead so it will
+        // still take a while to recover (as it needs to send padding until it reaches around
+        // 1000 kbps).
         this.newActiveStreams = ActiveStreams.kHi;
         newBitrates[0].maxBitrateKbps = 0;
         newBitrates[1].maxBitrateKbps = 1200;
@@ -404,11 +409,7 @@ export default class DefaultSimulcastUplinkPolicy implements SimulcastUplinkPoli
 
   forEachObserver(observerFunc: (observer: SimulcastUplinkObserver) => void): void {
     for (const observer of this.observerQueue) {
-      AsyncScheduler.nextTick(() => {
-        if (this.observerQueue.has(observer)) {
-          observerFunc(observer);
-        }
-      });
+      observerFunc(observer);
     }
   }
 }